Feature/hyunwoo

See merge request s12-webmobile4-sub1/S12P11D111!12

+++ Prepay/app/src/main/java/com/example/prepay/data/response/LikeTeamsReq.kt
@@ -1,6 +1,6 @@
 package com.example.prepay.data.response
 
 data class LikeTeamsReq(
-    val teamId: Long,
+    val teamId: Int,
     val checkLike: Boolean
 )
